\name{informinR-package}
\alias{informinR-package}
\alias{informinR}
\docType{package}
\title{
Tools to calculate informativeness scores
}
\description{
}
\details{
}
\author{
Ryan Raaum

Maintainer: Ryan Raaum <code@raaum.org>
}
\references{
}
\keyword{ package }
\seealso{
\code{\link[genotypeData:genotypeData-package]{genotypeData}}
}
\examples{
}

