\name{pcaContribution}
\alias{pcaContribution}
\title{pcaContribution}
\description{
}
\usage{
}
\details{
Calculate the PCA Contribution score for genotype data.
}
\value{
Returns a vector with the scores
}
\references{
}
\author{Ryan Raaum}
\examples{
## gd <- pcaContribution(gd)
}
\keyword{data}
